# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Book Search",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/organization/plans/{plan_id}/imports": {
      "post": {
        "tags": [
          "organization"
        ],
        "summary": "Start Import",
        "operationId": "start_import_api_organization_plans__plan_id__imports_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "organization"
        ],
        "summary": "Plan Imports",
        "operationId": "plan_imports_api_organization_plans__plan_id__imports_get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunView"
                  },
                  "title": "Response Plan Imports Api Organization Plans  Plan Id  Imports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organization/imports/{run_id}/entries/{entry_id}/cancel": {
      "post": {
        "tags": [
          "organization"
        ],
        "summary": "Cancel Entry",
        "operationId": "cancel_entry_api_organization_imports__run_id__entries__entry_id__cancel_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organization/imports/{run_id}": {
      "get": {
        "tags": [
          "organization"
        ],
        "summary": "Import Run",
        "operationId": "import_run_api_organization_imports__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organization/imports/{run_id}/entries/{entry_id}/retry": {
      "post": {
        "tags": [
          "organization"
        ],
        "summary": "Retry Entry",
        "operationId": "retry_entry_api_organization_imports__run_id__entries__entry_id__retry_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoverExportView": {
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "prepared",
              "unavailable"
            ],
            "title": "State"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "backend_selected": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backend Selected"
          },
          "unchanged": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unchanged"
          }
        },
        "type": "object",
        "required": [
          "state",
          "message"
        ],
        "title": "CoverExportView"
      },
...
      "DestinationChoice-Input": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "revision": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "revision"
        ],
        "title": "DestinationChoice"
      },
...
      "EntryView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "group_id": {
            "type": "string",
            "format": "uuid",
            "title": "Group Id"
          },
          "version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Version Id"
          },
          "destination_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Id"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "can_retry": {
            "type": "boolean",
            "title": "Can Retry",
            "default": false
          },
          "can_cancel": {
            "type": "boolean",
            "title": "Can Cancel",
            "default": false
          },
          "cover_export": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CoverExportView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "group_id",
          "version_id",
          "destination_id",
          "operation_id",
          "state",
          "message",
          "published_at",
          "confirmed_at",
          "asset_id"
        ],
        "title": "EntryView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportInput": {
        "properties": {
          "plan_revision": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Plan Revision"
          },
          "destinations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DestinationChoice-Input"
            },
            "propertyNames": {
              "enum": [
                "ebook",
                "audio"
              ]
            },
            "type": "object",
            "title": "Destinations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan_revision",
          "destinations"
        ],
        "title": "ImportInput"
      },
...
      "RunView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "plan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Plan Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EntryView"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "id",
          "plan_id",
          "created_at",
          "entries"
        ],
        "title": "RunView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}